Add files via upload
This is just a beta version of the app. I'm still working on it. There will be released version with all the features I wanted originally.
Pre-release version can open ".xls" file and convert it to modern ".xlsx" saving all the data except formatting to a new file with the same name preserving the original ".xls" file. The program won't overwrite original ".xls" if they present. Since this app relies on pandas, ".xlsx" files won't be copied  and will be used as is.
TO DO:
1. Improve UI/UX;
2. make "extractWindow" and it's features;
3. add comments to the code and translate anything in Russian to the English for better readability;
4. handle hard coded path to initial dir and make windows and linux compatible;
5. catch errors and improve notifications;
6. handle empty or None types so they won't crash the program;
7. handle multiple directories if files located in hierarchy;
8. release 1.0 version;
9. commit README file.

P.s, If you want to build it: (LINUX)
git clone https://github.com/Den0zavr/for-mum.git
cd for-mum
pyinstaller --onefile main.py
cd dist
wine ./main

And do not forget to copy or move main.exe to the actual directory with your files. Better have all your files in it for now.